+++ CHANGELOG.md
@@ -30,7 +30,11 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ## main
 
-### Breaking Changes
+### Breaking changes
+
+* Rename `BreadcrumbComponent` to `Beta::Breadcrumbs`.
+
+    *Joel Hawksley*
 
 * Split `UnderlineNavComponent` into `Alpha::UnderlineNav` and `Alpha::UnderlinePanels`.